// ui/src/app/shared/model/UpdateJobLabelsResponse.ts
/**
 * Job Monitor Service
 * Job Monitor API for interacting with asynchronous batch jobs and workflows.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * Response to an update labels request
 */
export interface UpdateJobLabelsResponse {
    /**
     * All job labels, after the update has been applied.
     */
    labels?: any;

}